Feat: added safety prompt before deletion

+++ src/core/delete_actions.rs
@@ -1,4 +1,5 @@
 use std::fs;
+use crate::cli::prompt;
 use crate::core::dates_names::DateNameForFile;
 use crate::prelude::*;
 use crate::core::file_access;
@@ -9,38 +10,58 @@ use super::date_models::units_validated::{ValidatedDate, ValidatedYear};
 use super::dates_names::yearly_name::YearlyName;
 use super::dates_names::{DailyName, MonthlyName};
 
-/// # Returns
-/// If true is returned then a daily journal was deleted. If false then there was no daily
-/// journal to be deleted
-/// ## Errors
-/// - could not even determine if a journal exits or could delte the found journal for some reason
-/// for example permission
-pub fn delete_day_journal(to_delete: ValidatedDate, option: &AppOptions) -> AppResult<bool> {
-    let daily_name: DailyName = to_delete.into();
-    delete_given_journal(daily_name, option)
-}
-/// # Returns
-/// If true is returned then a monthly journal was deleted. If false then there was no montly
-/// journal to be deleted
-/// ## Errors
-/// - could not even determine if a journal exits or could delte the found journal for some reason
-/// for example permission
-pub fn delete_month_journal(to_delete: OpenByMonthInYear, option: &AppOptions) -> AppResult<bool> {
-    let daily_name: MonthlyName = MonthlyName::from_month_in_year(&to_delete)?;
-    delete_given_journal(daily_name, option)
-}
-/// # Returns
-/// If true is returned then a yearly journal was deleted. If false then there was no yearly
-/// journal to be deleted
-/// ## Errors
-/// - could not even determine if a journal exits or could delte the found journal for some reason
-/// for example permission
-pub fn delete_year_journal(to_delete: ValidatedYear, option: &AppOptions) -> AppResult<bool> {
-    let daily_name: YearlyName = YearlyName::new(to_delete);
-    delete_given_journal(daily_name, option)
+pub enum DeletionResult {
+    /// There was journal to be found for deletion
+    NoJournalFound,
+    /// Use did cancel the deletion when asked for confirmation
+    NoConfirmation,
+    /// User confirmed and the journal was deleted
+    Deleted,
 }
 
-fn delete_given_journal<T>(journal: T, option: &AppOptions) -> AppResult<bool>
+/// ## Errors
+/// - could not even determine if a journal exits or could delte the found journal for some reason
+/// for example permission
+pub fn delete_day_journal(
+    to_delete: ValidatedDate,
+    option: &AppOptions,
+) -> AppResult<DeletionResult> {
+    let daily_name: DailyName = to_delete.into();
+    delete_given_journal(daily_name, option, || {
+        prompt::ask_for_confirmation("you want to delete the chosen daily journal ?")
+    })
+}
+/// ## Errors
+/// - could not even determine if a journal exits or could delte the found journal for some reason
+/// for example permission
+pub fn delete_month_journal(
+    to_delete: OpenByMonthInYear,
+    option: &AppOptions,
+) -> AppResult<DeletionResult> {
+    let daily_name: MonthlyName = MonthlyName::from_month_in_year(&to_delete)?;
+    delete_given_journal(daily_name, option, || {
+        prompt::ask_for_confirmation("you want to delete the chosen monthly journal ?")
+    })
+}
+
+/// ## Errors
+/// - could not even determine if a journal exits or could delte the found journal for some reason
+/// for example permission
+pub fn delete_year_journal(
+    to_delete: ValidatedYear,
+    option: &AppOptions,
+) -> AppResult<DeletionResult> {
+    let daily_name: YearlyName = YearlyName::new(to_delete);
+    delete_given_journal(daily_name, option, || {
+        prompt::ask_for_confirmation("you want to delete the chosen yearly journal ?")
+    })
+}
+
+fn delete_given_journal<T>(
+    journal: T,
+    option: &AppOptions,
+    on_confirmation: impl Fn() -> AppResult<bool>,
+) -> AppResult<DeletionResult>
 where
     T: DateNameForFile,
 {
@@ -50,14 +71,19 @@ where
         Ok(does_exits) => {
             if does_exits {
                 if to_open.is_file() {
-                    fs::remove_file(to_open)?;
-                    Ok(true)
+                    let wants_to_delete = on_confirmation()?;
+                    if wants_to_delete {
+                        fs::remove_file(to_open)?;
+                        Ok(DeletionResult::Deleted)
+                    } else {
+                        Ok(DeletionResult::NoConfirmation)
+                    }
                 } else {
                     bail!("Critical: found directory where file as journal was expected to be deleted")
                 }
             } else {
                 // no journal found to delete
-                Ok(false)
+                Ok(DeletionResult::NoJournalFound)
             }
         }
         Err(error) => Err(AppError::new(error).context(
